requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.3.0
requests==2.31.0
sqlalchemy==2.0.23
pandas==2.2.0
numpy==1.24.3
plotly==5.17.0
python-dotenv==1.0.0
scikit-learn==1.3.0
matplotlib==3.8.0